pyproject.toml
[project]
name = "reflux-bot"
version = "0.1.0"
description = "Telegram bot for reflux tracking (meals/symptoms/meds/morning checks) with SQLite storage."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0,<22.0",
  "SQLAlchemy>=2.0,<3.0",
  "python-dotenv>=1.0,<2.0",
]

[tool.ruff]
line-length = 100


